+++ daemons/hw_mouseremote.c
@@ -220,7 +220,7 @@ char *mouseremote_rec(struct ir_remote *remotes)
 		    (char)(b[2]) == 0x3f && ((char)(b[2]) & 0x07)) {
 			code=(ir_code) (char)(b[1]) | 
 			               (((char)(b[0]) & 0x03)<<6);
-			LOGPRINTF(1,"result %llx", (unsigned long long) code);
+			LOGPRINTF(1,"result %llx", (__u64) code);
 			m=decode_all(remotes);
 			return(m);
 		}
@@ -237,7 +237,7 @@ char *mouseremote_rec(struct ir_remote *remotes)
 				return(NULL);
 			}
 			code = (ir_code)b[1];
-			LOGPRINTF(1,"result %llx", (unsigned long long) code);
+			LOGPRINTF(1,"result %llx", (__u64) code);
 			m=decode_all(remotes);
 			return(m);
 		}
@@ -278,12 +278,12 @@ char *mouseremote_rec(struct ir_remote *remotes)
 	}
 	if (code != 0) {
 		code |= 0x0100;
-		LOGPRINTF(1,"result %llx", (unsigned long long) code);
+		LOGPRINTF(1,"result %llx", (__u64) code);
 		m=decode_all(remotes);
 		return(m);
 	} else if (dx == 0 && dy == 0) {
 		code = 0x0800 | stat;
-		LOGPRINTF(1,"result %llx", (unsigned long long) code);
+		LOGPRINTF(1,"result %llx", (__u64) code);
 		m=decode_all(remotes);
 		return(m);
 	}
