+++ drivers/lirc_dev/lirc_dev.h
@@ -330,10 +330,10 @@ static inline unsigned int lirc_buffer_write_n(struct lirc_buffer *buf,
 struct lirc_driver {
 	char name[40];
 	int minor;
-	unsigned long code_length;
+	__u32 code_length;
 	unsigned int buffer_size; /* in chunks holding one code each */
 	int sample_rate;
-	unsigned long features;
+	__u32 features;
 	void *data;
 	lirc_t min_timeout;
 	lirc_t max_timeout;
