+++ tools/smode2.c
@@ -89,7 +89,7 @@ void closescreen(void)
 int main(int argc, char **argv)
 {
 	int fd;
-	unsigned long mode;
+	__u32 mode;
 	lirc_t data;
 	lirc_t x1,y1,x2,y2;
 	int result;
